+++ examples/noise_handshake.rs
@@ -11,9 +11,8 @@ use tracing::*;
 use tracing_subscriber::filter::LevelFilter;
 
 use pea2pea::{
-    connections::ConnectionSide,
     protocols::{Handshaking, Reading, ReturnableConnection, Writing},
-    Node, NodeConfig, Pea2Pea,
+    ConnectionSide, Node, NodeConfig, Pea2Pea,
 };
 
 use std::{